chore: split tests into several modules
